Use less precise floats.

+++ src/math/expressions.rs
@@ -47,7 +47,7 @@ impl Expression {
         insert_side(&mut self.right_hand_side, to_move, insert_at_start);
     }
 
-    pub fn mul_both_sides(&mut self, by: f64) {
+    pub fn mul_both_sides(&mut self, by: f32) {
         mul_side(&mut self.left_hand_side, by);
         mul_side(&mut self.right_hand_side, by);
 
@@ -87,7 +87,7 @@ fn insert_side(side: &mut Vec<AbstVar>, var: AbstVar, start: bool) {
     }
 }
 
-fn mul_side(side: &mut Vec<AbstVar>, by: f64) {
+fn mul_side(side: &mut Vec<AbstVar>, by: f32) {
     for var in side.iter_mut() {
         match var {
             &mut AbstVar::Variable { ref mut coefficient, .. } => {
